+++ iceberg/iceberg-rest-server/build.gradle.kts
@@ -31,6 +31,7 @@ val icebergVersion: String = libs.versions.iceberg.get()
 val scalaCollectionCompatVersion: String = libs.versions.scala.collection.compat.get()
 
 dependencies {
+  implementation(project(":catalogs:bundled-catalog", configuration = "shadow"))
   implementation(project(":core"))
   implementation(project(":common"))
   implementation(project(":iceberg:iceberg-common"))
