+++ iceberg/iceberg-rest-server/src/test/java/org/apache/gravitino/iceberg/service/rest/TestIcebergNamespaceOperations.java
@@ -1,4 +1,4 @@
-package org.apache.gravitino.iceberg.service.rest;/*
+/*
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
  * distributed with this work for additional information
@@ -16,7 +16,7 @@ package org.apache.gravitino.iceberg.service.rest;/*
  * specific language governing permissions and limitations
  * under the License.
  */
-
+package org.apache.gravitino.iceberg.service.rest;
 
 import com.google.common.collect.ImmutableMap;
 import java.util.Arrays;
