+++ integration-test-common/src/test/java/org/apache/gravitino/integration/test/util/AbstractIT.java
@@ -49,10 +49,6 @@ import org.apache.gravitino.integration.test.container.MySQLContainer;
 import org.apache.gravitino.server.GravitinoServer;
 import org.apache.gravitino.server.ServerConfig;
 import org.apache.gravitino.server.web.JettyServerConfig;
-import org.apache.hc.client5.http.classic.methods.HttpGet;
-import org.apache.hc.client5.http.impl.classic.CloseableHttpClient;
-import org.apache.hc.client5.http.impl.classic.HttpClients;
-import org.apache.hc.core5.http.ClassicHttpResponse;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -253,7 +249,7 @@ public class AbstractIT {
       Awaitility.await()
           .atMost(60, TimeUnit.SECONDS)
           .pollInterval(1, TimeUnit.SECONDS)
-          .until(() -> isHttpServerUp(checkServerUrl));
+          .until(() -> HttpUtils.isHttpServerUp(checkServerUrl));
     }
 
     JettyServerConfig jettyServerConfig =
@@ -314,26 +310,6 @@ public class AbstractIT {
     } catch (IOException e) {
       LOG.warn("Can't get git commit id for:", e);
       return "";
-    }
-  }
-
-  /**
-   * Check if the http server is up, If http response status code is 200, then we're assuming the
-   * server is up. Or else we assume the server is not ready.
-   *
-   * <p>Note: The method will ignore the response body and only check the status code.
-   *
-   * @param testUrl A url that we want to test ignore the response body.
-   * @return true if the server is up, false otherwise.
-   */
-  public static boolean isHttpServerUp(String testUrl) {
-    try (CloseableHttpClient httpClient = HttpClients.createDefault()) {
-      HttpGet request = new HttpGet(testUrl);
-      ClassicHttpResponse response = httpClient.execute(request, a -> a);
-      return response.getCode() == 200;
-    } catch (Exception e) {
-      LOG.warn("Check gravitino server failed: url:{}, error message:{}", testUrl, e.getMessage());
-      return false;
     }
   }
 }
